requirements.txt
# =========================
# === Core Web Application ===
# =========================
fastapi>=0.110.0
fastapi-mcp>=0.1.5
uvicorn[standard]>=0.22
PyGithub>=2.3.0

# =========================
# === Google Docs & Sync  ===
# =========================
google-api-python-client>=2.70
google-auth>=2.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
httplib2>=0.22
markdownify>=1.1.0      # HTML to Markdown converter

# =========================
# === Markdown & Text Processing ===
# =========================
markdown>=3.4.1         # Standard Python Markdown parser (used for rendering .md)
python-frontmatter>=1.1.0  # For reading YAML frontmatter in markdown files
          
# =========================
# === Environment & Config ===
# =========================
python-dotenv>=1.0.0    # Loads environment variables from .env

# =========================
# === HTTP Clients & Utilities ===
# =========================
aiohttp>=3.8            # Async HTTP client
httpx>=0.28             # Modern async/sync HTTP client (recommended for FastAPI)
requests>=2.31          # Standard synchronous HTTP requests
requests-oauthlib>=2.0.0
aiohappyeyeballs>=2.6   # Fast DNS resolution for aiohttp
aiofiles>=23.1.0        # Async file I/O

# =========================
# === AI, LLMs, OpenAI, LangChain ===
# =========================
openai>=1.30.5
transformers>=4.30
langchain>=0.2
langchain-community>=0.2
langchain-core>=0.2
langchain-openai>=0.1

# =========================
# === Semantic Search & Embeddings ===
# =========================
llama-index>=0.12.43
tiktoken==0.7.0
tree_sitter>=0.20.4
tree_sitter_language_pack>=0.0.1

# =========================
# === Monitoring & Diagnostics ===
# =========================
psutil>=5.9

# =========================
# === Authentication & Security (JWT, sessions, etc) ===
# =========================
itsdangerous>=2.1
pyjwt[crypto]>=2.8

# =========================
# === Structured Data & Validation ===
# =========================
pydantic>=2.11.7

# =========================
# === Async Utilities ===
# =========================
anyio>=3.7

# =========================
# === Databases ===
# =========================
sqlalchemy>=2.0
neo4j>=5.16.0
psycopg2-binary>=2.9


# =========================
# === Testing ===
# =========================
pytest>=7.0
pytest-asyncio>=0.21

# =========================
# === Linting & Formatting ===
# =========================
black>=24.0

# =========================
# === Observability: OpenTelemetry Tracing ===
# =========================
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-http>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-asgi>=0.46b0  # Needed for FastAPI, covers all ASGI middleware
opentelemetry-instrumentation-httpx>=0.46b0   # traces outbound HTTPX calls
opentelemetry-instrumentation-aiohttp-client>=0.46b0  # traces outbound aiohttp calls (if used)
# (optional convenience meta-package; installs both grpc+http OTLP exporters)
# opentelemetry-exporter-otlp>=1.25.0

# =========================
# === Automation        ===
# =========================
playwright==1.42.0

# === File/JSON Parsing (used in scraper logic) ===
python-dateutil>=2.9

# =========================
# === (Optional) Markdown/HTML Extensions ===
# =========================
 bleach>=6.1.0  # Sanitize HTML if converting markdown to HTML
# mistune       # Advanced/fast markdown parser with plugins
# pypandoc      # Universal document converter for .md/.docx/.html

# =========================
# === Notes ===
# =========================
# - All packages are pinned to major versions; you may want to pin more tightly in production.
# - For frontend markdown (JS/React), see your package.json (not this file).
# - If you don’t use a feature, you can comment out its dependencies to speed up installs.

